MNT move PolynomialFeatures from _data.py to _polynomial.py (#19611)
